+++ Snoop/InjectorLauncherManager.cs
@@ -18,7 +18,10 @@ namespace Snoop
     /// </summary>
     public static class InjectorLauncherManager
     {
-        private static string GetSuffix(ProcessInfo processInfo)
+        // This method has to work without exceptions.
+        // If we get the architecture wrong here the InjectorLauncher will fix this by starting a secondary instance.
+        // IsProcess64BitWithoutException will return false in case of a failure, this way we ensure that the x86 InjectorLauncher is run, which should be able to run on every system.
+        private static string GetArchitectureSuffix(ProcessInfo processInfo)
         {
             var bitness = processInfo.IsProcess64Bit
                 ? "x64"
@@ -43,7 +46,7 @@ namespace Snoop
             {
                 var location = Assembly.GetExecutingAssembly().Location;
                 var directory = Path.GetDirectoryName(location) ?? string.Empty;
-                var injectorLauncherExe = Path.Combine(directory, $"Snoop.InjectorLauncher.{GetSuffix(processInfo)}.exe");
+                var injectorLauncherExe = Path.Combine(directory, $"Snoop.InjectorLauncher.{GetArchitectureSuffix(processInfo)}.exe");
 
                 if (File.Exists(injectorLauncherExe) == false)
                 {